requirements.txt
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
pyinstaller==5.13.0
pillow==10.0.0